refactor: Rename filterId fun
